Space Invaders Game
This game targets a young age group of seven to fifteen years old. This
game will be situated in a kiosk PC. The game is quite user friendly and
it is not difficult to play and win. It also has a space theme and the
graphics are adequate for the targeted age group.